+++ src/recipes.js
@@ -1,0 +1,11 @@
+const filterRecipes = (data, filterTerm) => {
+  let filteredRecipes = data.filter((recipe) => {
+     return recipe.tags.includes(filterTerm) || recipe.name === filterTerm
+  })
+  if(filteredRecipes.length === 0) {
+    return 'Sorry, no matching results!'
+  }
+  return filteredRecipes
+}
+
+export {filterRecipes}
